# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Talyn Gold Trading API",
        "description": "API for gold trading platform",
        "contact": {
            "name": "Talyn Support",
            "email": "info@talyn.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get the authenticated user's information",
                "operationId": "dd48b9887ac0e98fac70ebd3da3eb645",
                "responses": {
                    "200": {
                        "description": "User information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Shahriyar"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "shahriyar@email.com"
                                        },
                                        "gold_balance": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 8.517
                                        },
                                        "cash_balance": {
                                            "type": "integer",
                                            "example": 100000000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/balance": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get the authenticated user's balance",
                "operationId": "f3677d2581f726411270c109de01e66a",
                "responses": {
                    "200": {
                        "description": "User balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "gold_balance": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 8.517
                                        },
                                        "gold_balance_formatted": {
                                            "type": "string",
                                            "example": "8.517 گرم"
                                        },
                                        "cash_balance": {
                                            "type": "integer",
                                            "example": 100000000
                                        },
                                        "cash_balance_rial": {
                                            "type": "string",
                                            "example": "100,000,000 ریال"
                                        },
                                        "cash_balance_toman": {
                                            "type": "string",
                                            "example": "10,000,000 تومان"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}